try to move stuff out of objectloader2
